set up new data version of nl interface
