+++ server/__init__.py
@@ -179,6 +179,9 @@ def register_routes_common(app):
   from server.routes.translator import api as translator_api
   app.register_blueprint(translator_api.bp)
 
+  from server.routes.nl import api as nl_api
+  app.register_blueprint(nl_api.bp)
+
   from server.routes.shared_api import choropleth as shared_choropleth
   app.register_blueprint(shared_choropleth.bp)
