+++ static/webpack.config.js
@@ -105,6 +105,9 @@ const config = {
       __dirname + "/js/apps/sustainability/main.ts",
       __dirname + "/css/sustainability.scss",
     ],
+    nl_interface_data: [
+      __dirname + "/js/apps/nl_interface/data_app/main.ts"
+    ]
   },
   output: {
     path: path.resolve(__dirname, "../") + "/server/dist",
